# perf: side effects when there aren't too many optimizable dependencies