--- a/crates/rspack_core/src/exports_info.rs
+++ b/crates/rspack_core/src/exports_info.rs
@@ -1307,9 +1307,9 @@ impl ExportInfo {
     false
   }
 
-  pub fn move_target<'a>(
-    &'a self,
-    mg: &'a mut ModuleGraph<'a>,
+  pub fn move_target(
+    &self,
+    mg: &mut ModuleGraph,
     resolve_filter: ResolveFilterFnTy,
     update_original_connection: UpdateOriginalFunctionTy,
   ) -> Option<ResolvedExportInfoTarget> {
@@ -1896,7 +1896,7 @@ impl MaybeDynamicTargetExportInfo {
   }
 }
 
-pub type ResolveFilterFnTy = Rc<dyn Fn(&ResolvedExportInfoTarget, &ModuleGraph) -> bool>;
+pub type ResolveFilterFnTy<'a> = Rc<dyn Fn(&ResolvedExportInfoTarget, &ModuleGraph) -> bool + 'a>;
 
 fn resolve_target(
   input_target: Option<UnResolvedExportInfoTarget>,